Fix VerticaPyLab start bug

+++ Makefile
@@ -110,9 +110,17 @@ vsql: ## Run a basic sanity test (optional -DQUERY="select 'whatever')
 .PHONY: verticapylab-start
 verticapylab-start: etc/VerticaPyLab.conf ## Start a jupyterlab
 	@source etc/VerticaPyLab.conf; \
+	if [[ $${TEST_MODE^^} == "YES" ]] ; then \
+		VERTICAPYLAB_IMG_VERSION=$(VERSION); \
+	fi; \
 	if (($$(docker ps --no-trunc -q -f NAME="$$VERTICAPYLAB_CONTAINER_NAME" | wc -l)==0)); then \
 	    if [[ -z $$(docker image ls -q "opentext/$$VERTICAPYLAB_IMG:$$VERTICAPYLAB_IMG_VERSION" 2>&1) ]]; then \
-	      $(MAKE) verticapylab-install || exit 1; \
+		  if [[ $${TEST_MODE^^} == "YES" ]] ; then \
+		    echo "Building image opentext/$$VERTICAPYLAB_IMG:$$VERTICAPYLAB_IMG_VERSION"; \
+			TEST_MODE=yes $(MAKE) verticapylab-build; \
+		  else \
+			$(MAKE) verticapylab-install || exit 1; \
+		  fi; \
 	    fi; \
 	    docker container rm "$$VERTICAPYLAB_CONTAINER_NAME" >/dev/null 2>&1; \
 	    bin/verticapylab; \
